+++ src/main/java/chzzk/grassdiary/auth/common/AuthenticatedMember.java
@@ -1,0 +1,11 @@
+package chzzk.grassdiary.auth.common;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(value = RetentionPolicy.RUNTIME)
+@Target(value = ElementType.PARAMETER)
+public @interface AuthenticatedMember {
+}
