+++ src/main/java/chzzk/grassdiary/auth/controller/OAuthTestController.java
@@ -1,0 +1,16 @@
+package chzzk.grassdiary.auth.controller;
+
+import chzzk.grassdiary.auth.common.AuthenticatedMember;
+import chzzk.grassdiary.auth.service.dto.AuthMemberPayload;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class OAuthTestController {
+
+    @GetMapping("/test")
+    public String test(@AuthenticatedMember AuthMemberPayload authMemberPayload) {
+        System.out.println("member = " + authMemberPayload);
+        return "hello!";
+    }
+}
