+++ src/main/java/chzzk/grassdiary/auth/exception/AuthenticationException.java
@@ -1,0 +1,7 @@
+package chzzk.grassdiary.auth.exception;
+
+public class AuthenticationException extends RuntimeException {
+    public AuthenticationException(String message) {
+        super(message);
+    }
+}
