+++ src/main/java/chzzk/grassdiary/auth/exception/JwtException.java
@@ -1,0 +1,11 @@
+package chzzk.grassdiary.auth.exception;
+
+public class JwtException extends RuntimeException {
+    public JwtException(String message) {
+        super(message);
+    }
+
+    public JwtException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
