+++ src/main/java/chzzk/grassdiary/auth/service/dto/GoogleOAuthToken.java
@@ -1,0 +1,9 @@
+package chzzk.grassdiary.auth.service.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record GoogleOAuthToken(
+        @JsonProperty("access_token")
+        String accessToken
+) {
+}
