+++ src/main/java/chzzk/grassdiary/auth/service/dto/JWTTokenResponse.java
@@ -1,0 +1,4 @@
+package chzzk.grassdiary.auth.service.dto;
+
+public record JWTTokenResponse(String accessToken) {
+}
