+++ src/main/java/chzzk/grassdiary/auth/util/OAuthUriGenerator.java
@@ -1,0 +1,6 @@
+package chzzk.grassdiary.auth.util;
+
+
+public interface OAuthUriGenerator {
+    String generateUrl();
+}
