--- src/main/java/chzzk/grassdiary/domain/auth/common/AuthenticatedMember.java
@@ -1,11 +1,0 @@
-package chzzk.grassdiary.domain.auth.common;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(value = RetentionPolicy.RUNTIME)
-@Target(value = ElementType.PARAMETER)
-public @interface AuthenticatedMember {
-}
