--- src/main/java/chzzk/grassdiary/domain/auth/exception/AuthenticationException.java
@@ -1,7 +1,0 @@
-package chzzk.grassdiary.domain.auth.exception;
-
-public class AuthenticationException extends RuntimeException {
-    public AuthenticationException(String message) {
-        super(message);
-    }
-}
