--- src/main/java/chzzk/grassdiary/domain/auth/exception/JwtException.java
@@ -1,11 +1,0 @@
-package chzzk.grassdiary.domain.auth.exception;
-
-public class JwtException extends RuntimeException {
-    public JwtException(String message) {
-        super(message);
-    }
-
-    public JwtException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
