--- src/main/java/chzzk/grassdiary/domain/auth/jwt/JwtTokenExtractor.java
@@ -1,25 +1,0 @@
-package chzzk.grassdiary.domain.auth.jwt;
-
-import chzzk.grassdiary.domain.auth.exception.AuthenticationException;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpHeaders;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-
-@Slf4j
-@Component
-public class JwtTokenExtractor {
-    private static final String PREFIX_BEARER = "Bearer ";
-    private static final String ACCESS_TOKEN_HEADER = HttpHeaders.AUTHORIZATION;
-
-    public String extractAccessToken(HttpServletRequest request) {
-        String accessToken = request.getHeader(ACCESS_TOKEN_HEADER);
-
-        if (StringUtils.hasText(accessToken) && accessToken.startsWith(PREFIX_BEARER)) {
-            return accessToken.substring(PREFIX_BEARER.length());
-        }
-        String logMessage = "[인증 실패] 액세스 토큰 추출 실패 - 토큰: " + accessToken;
-        throw new AuthenticationException(logMessage);
-    }
-}
