--- src/main/java/chzzk/grassdiary/domain/auth/service/dto/GoogleOAuthToken.java
@@ -1,9 +1,0 @@
-package chzzk.grassdiary.domain.auth.service.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record GoogleOAuthToken(
-        @JsonProperty("access_token")
-        String accessToken
-) {
-}
