--- src/main/java/chzzk/grassdiary/domain/auth/service/dto/GoogleUserInfo.java
@@ -1,17 +1,0 @@
-package chzzk.grassdiary.domain.auth.service.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record GoogleUserInfo(
-        String id,
-
-        @JsonProperty("email")
-        String email,
-
-        @JsonProperty("given_name")
-        String nickname,
-
-        @JsonProperty("picture")
-        String picture
-) {
-}
