--- src/main/java/chzzk/grassdiary/domain/auth/service/dto/JWTTokenResponse.java
@@ -1,4 +1,0 @@
-package chzzk.grassdiary.domain.auth.service.dto;
-
-public record JWTTokenResponse(String accessToken) {
-}
