--- src/main/java/chzzk/grassdiary/domain/auth/util/OAuthUriGenerator.java
@@ -1,6 +1,0 @@
-package chzzk.grassdiary.domain.auth.util;
-
-
-public interface OAuthUriGenerator {
-    String generateUrl();
-}
