--- src/main/java/chzzk/grassdiary/domain/global/presentation/ErrorResponse.java
@@ -1,4 +1,0 @@
-package chzzk.grassdiary.domain.global.presentation;
-
-public record ErrorResponse(String error) {
-}
