+++ src/main/java/chzzk/grassdiary/global/presentation/ErrorResponse.java
@@ -1,0 +1,4 @@
+package chzzk.grassdiary.global.presentation;
+
+public record ErrorResponse(String error) {
+}
