+++ src/main/java/chzzk/grassdiary/global/presentation/GlobalExceptionHandler.java
@@ -1,0 +1,22 @@
+package chzzk.grassdiary.global.presentation;
+
+import chzzk.grassdiary.auth.exception.AuthenticationException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(value = AuthenticationException.class)
+    public ResponseEntity<ErrorResponse> handleAuthenticationException(AuthenticationException exception) {
+        log.warn("UnAuthorizationException | " + exception.getMessage());
+
+        return ResponseEntity
+                .status(HttpStatus.UNAUTHORIZED)
+                .body(new ErrorResponse("인증을 실패했습니다."));
+    }
+}
